moved side panel div into separate component
