+++ src/client/utils.js
@@ -1,0 +1,3 @@
+export const openLink = (url) => {
+  window.open(url,'_blank', 'noopener,noreferrer');
+};
